+++ UiApp/premake5.lua
@@ -32,10 +32,18 @@ project "UiApp"
 			"UI_PLATFORM_WINDOWS"
 		}
 
+	filter "system:linux"
+		cppdialect "C++17"
+		staticruntime "On"
+		systemversion "latest"
+
+        links { "GLFW", "Xrandr", "Xi", "GLU", "GL", "X11", "dl", "pthread", "stdc++fs" }
+        defines { "_X11", "UI_PLATFORM_LINUX" }
+
 	filter "configurations:Debug"
-		defines "HZ_DEBUG"
+		defines "UI_DEBUG"
 		symbols "On"
 
 	filter "configurations:Release"
-		defines "HZ_RELEASE"
+		defines "UI_RELEASE"
 		optimize "On"