+++ UiEngine/src/Platform/Windows/WindowsWindow.h
@@ -1,0 +1,38 @@
+#pragma once
+
+#include <GLFW/glfw3.h>
+#include <iostream>
+#include <imgui.h>
+#include <backends/imgui_impl_opengl3.h>
+#include <backends/imgui_impl_glfw.h>
+
+#include "UiEngine/Window.h"
+
+namespace UiEngine
+{
+	class WindowsWindow : public Window
+	{
+	public:
+		WindowsWindow(const WindowProps& props);
+		virtual ~WindowsWindow();
+
+		void OnUpdate() override;
+		inline unsigned int GetWidth() const override { return m_Data.Width; }
+		inline unsigned int GetHeigth() const override { return m_Data.Height; }
+
+	private:
+		virtual void Init(const WindowProps& props);
+		virtual void Shutdown();
+
+	private:
+		GLFWwindow* m_Window;
+		struct WindowData
+		{
+			std::string Title;
+			unsigned int Width, Height;
+		};
+		WindowData m_Data;
+	};
+}
+
+