+++ UiEngine/src/UiEngine/EntryPoint.h
@@ -1,14 +1,25 @@
 #pragma once
 
+#include "UiEngine/Window.h"
+#include "Platform/Windows/WindowsWindow.h"
+#include "Platform/Linux/LinuxWindow.h"
+
 #ifdef UI_PLATFORM_WINDOWS
-
-extern UiEngine::Application* UiEngine::CreateApplication();
-
-int main(int argc, char** argv)
+namespace UiEngine
 {
-	auto app = UiEngine::CreateApplication();
-	app->Run();
-	delete app;
+	Window* Window::Create(const WindowProps& props)
+	{
+		std::cout << "Windows" << std::endl;
+		return new WindowsWindow(props);
+	}
 }
-
+#elif UI_PLATFORM_LINUX
+namespace UiEngine
+{
+	Window* Window::Create(const WindowProps& props)
+	{
+		std::cout << "Linux" << std::endl;
+		return new LinuxWindow(props);
+	}
+}
 #endif