+++ changelog.md
@@ -3,3 +3,8 @@
 Changelog added.
 
 ## 29.04.2022 - Github Actions are added
+## 28.07.2022 - Multi Platform Support
+* Imgui added as submodule
+* GLFW added as submodule
+* Platform spesific window class generated with inheritance
+* Linux-windows project generation and run shortcuts added into scripts folder